# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Uganda Diaspora Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /embassies:
    get:
      operationId: listEmbassies
      tags: [embassies]
      summary: List embassies
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: region, in: query, schema: { type: string } }
        - { name: continent, in: query, schema: { type: string } }
      responses:
        "200":
          description: Embassy list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbassyListResponse"
    post:
      operationId: createEmbassy
      tags: [embassies]
      summary: Create embassy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmbassyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Embassy"
  /embassies/{id}:
    get:
      operationId: getEmbassy
      tags: [embassies]
      summary: Get embassy
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Embassy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Embassy"
        "404":
          description: Not found
    patch:
      operationId: updateEmbassy
      tags: [embassies]
      summary: Update embassy
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmbassyUpdate"
      responses:
        "200":
          description: Updated embassy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Embassy"
    delete:
      operationId: deleteEmbassy
      tags: [embassies]
      summary: Delete embassy
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # TOURISM
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: upcoming, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # POSTS
...
  /mdas:
    get:
      operationId: listMdas
      tags: [mdas]
      summary: List government MDAs
      responses:
        "200":
          description: MDA list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mda"
    post:
      operationId: createMda
      tags: [mdas]
      summary: Create MDA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MdaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mda"
  /mdas/{id}:
    get:
      operationId: getMda
      tags: [mdas]
      summary: Get MDA
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: MDA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mda"
        "404":
          description: Not found
    patch:
      operationId: updateMda
      tags: [mdas]
      summary: Update MDA
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MdaUpdate"
      responses:
        "200":
          description: Updated MDA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mda"
    delete:
      operationId: deleteMda
      tags: [mdas]
      summary: Delete MDA
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # OPPORTUNITIES
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
    RegisterInput:
      type: object
      required: [email, password, fullName]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        fullName: { type: string }
        country: { type: ["string", "null"] }
        role: { type: ["string", "null"] }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, email, fullName, role, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        fullName: { type: string }
        role: { type: string }
        country: { type: ["string", "null"] }
        profession: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        isVerified: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    Embassy:
      type: object
      required: [id, country, city, createdAt]
      properties:
        id: { type: integer }
        country: { type: string }
        city: { type: string }
        continent: { type: ["string", "null"] }
        region: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        flagUrl: { type: ["string", "null"] }
        ambassadorName: { type: ["string", "null"] }
        ambassadorImageUrl: { type: ["string", "null"] }
        officeHours: { type: ["string", "null"] }
        servicesOffered: { type: ["string", "null"] }
        emergencyContact: { type: ["string", "null"] }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    EmbassyInput:
      type: object
      required: [country, city]
      properties:
        country: { type: string }
        city: { type: string }
        continent: { type: string }
        region: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        website: { type: string }
        imageUrl: { type: string }
        flagUrl: { type: string }
        ambassadorName: { type: string }
        ambassadorImageUrl: { type: string }
        officeHours: { type: string }
        servicesOffered: { type: string }
        emergencyContact: { type: string }
        latitude: { type: number }
        longitude: { type: number }
    EmbassyUpdate:
      type: object
      properties:
        country: { type: string }
        city: { type: string }
        continent: { type: string }
        region: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        website: { type: string }
        imageUrl: { type: string }
        ambassadorName: { type: string }
        officeHours: { type: string }
        servicesOffered: { type: string }
        emergencyContact: { type: string }
        isActive: { type: boolean }
    EmbassyListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Embassy"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # TOURISM
...
    Event:
      type: object
      required: [id, title, startDate, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        startDate: { type: string, format: date-time }
        endDate: { type: ["string", "null"] }
        registrationUrl: { type: ["string", "null"] }
        isVirtual: { type: boolean }
        isPublished: { type: boolean }
        registrationCount: { type: integer }
        createdAt: { type: string, format: date-time }
    EventInput:
      type: object
      required: [title, startDate]
      properties:
        title: { type: string }
        description: { type: string }
        location: { type: string }
        imageUrl: { type: string }
        category: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        registrationUrl: { type: string }
        isVirtual: { type: boolean }
        isPublished: { type: boolean }
    EventUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        location: { type: string }
        imageUrl: { type: string }
        category: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        isVirtual: { type: boolean }
        isPublished: { type: boolean }
    EventListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # POSTS
...
    Mda:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    MdaInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        logoUrl: { type: string }
        website: { type: string }
        category: { type: string }
    MdaUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        logoUrl: { type: string }
        website: { type: string }
        category: { type: string }
        isActive: { type: boolean }

    # OPPORTUNITIES
...