# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Uganda Diaspora Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  fullName: string;
  role: string;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  profession?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isVerified?: boolean;
  isActive?: boolean;
  createdAt: Date;
}